pyproject.toml
[tool.poetry]
name = "simple_short_links"
version = "0.1.0"
description = ""
authors = ["Yaroslav Pisarev <yaricp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
hashids = "^1.3.1"
pydantic = {extras = ["email"], version = "^1.9.0"}
SQLAlchemy = "^1.4.32"
alembic = "^1.7.7"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"
databases = {extras = ["postgresql"], version = "^0.5.5"}
psycopg2 = "^2.9.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
fastapi-utils = "^0.2.1"
flake8 = "^4.0.1"
pytest = "^7.1.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
    in-project=true
    create=true
